redis/python/requirements.txt
Flask==2.3.3
werkzeug==2.3.7
redis==4.5.0
python-dotenv==1.0.0